Split derive macro "Compressed" into "Compress" and "Decompress"
