+++ fuel-derive/src/compression/compress.rs
@@ -1,134 +1,18 @@
 use proc_macro2::{
     Span,
     TokenStream as TokenStream2,
-    TokenTree as TokenTree2,
 };
 use quote::{
     format_ident,
     quote,
 };
 
-use syn::parse::{
-    Parse,
-    ParseStream,
+use crate::helpers::where_clause_push;
+
+use super::attribute::{
+    FieldAttrs,
+    StructureAttrs,
 };
-
-const ATTR: &str = "da_compress";
-
-/// Structure (struct or enum) attributes
-#[derive(Debug)]
-pub enum StructureAttrs {
-    /// Insert bounds for a generic type
-    /// `#[da_compress(bound(Type))]`
-    Bound(Vec<String>),
-    /// Discard generic parameter
-    /// `#[da_compress(discard(Type))]`
-    Discard(Vec<String>),
-}
-impl Parse for StructureAttrs {
-    fn parse(input: ParseStream) -> syn::Result<Self> {
-        if let Ok(ml) = input.parse::<syn::MetaList>() {
-            if ml.path.segments.len() == 1 {
-                match ml.path.segments[0].ident.to_string().as_str() {
-                    "bound" => {
-                        let mut bound = Vec::new();
-                        for item in ml.tokens {
-                            match item {
-                                TokenTree2::Ident(ident) => {
-                                    bound.push(ident.to_string());
-                                }
-                                other => {
-                                    return Err(syn::Error::new_spanned(
-                                        other,
-                                        "Expected generic (type) name",
-                                    ))
-                                }
-                            }
-                        }
-                        return Ok(Self::Bound(bound));
-                    }
-                    "discard" => {
-                        let mut discard = Vec::new();
-                        for item in ml.tokens {
-                            match item {
-                                TokenTree2::Ident(ident) => {
-                                    discard.push(ident.to_string());
-                                }
-                                other => {
-                                    return Err(syn::Error::new_spanned(
-                                        other,
-                                        "Expected generic (type) name",
-                                    ))
-                                }
-                            }
-                        }
-                        return Ok(Self::Discard(discard));
-                    }
-                    _ => {}
-                }
-            }
-        }
-        Err(syn::Error::new_spanned(
-            input.parse::<syn::Ident>()?,
-            "Expected `bound` or `discard`",
-        ))
-    }
-}
-impl StructureAttrs {
-    pub fn parse(attrs: &[syn::Attribute]) -> syn::Result<Vec<Self>> {
-        let mut result = Vec::new();
-        for attr in attrs {
-            if attr.style != syn::AttrStyle::Outer {
-                continue;
-            }
-
-            if let syn::Meta::List(ml) = &attr.meta {
-                if ml.path.segments.len() == 1 && ml.path.segments[0].ident == ATTR {
-                    result.push(syn::parse2::<StructureAttrs>(ml.tokens.clone())?);
-                }
-            }
-        }
-
-        Ok(result)
-    }
-}
-
-/// Field attributes
-pub enum FieldAttrs {
-    /// Skipped when compressing, and must be reconstructed when decompressing.
-    /// `#[da_compress(skip)]`
-    Skip,
-    /// Compresseded recursively.
-    Normal,
-}
-impl FieldAttrs {
-    pub fn parse(attrs: &[syn::Attribute]) -> Self {
-        let mut result = Self::Normal;
-        for attr in attrs {
-            if attr.style != syn::AttrStyle::Outer {
-                continue;
-            }
-
-            if let syn::Meta::List(ml) = &attr.meta {
-                if ml.path.segments.len() == 1 && ml.path.segments[0].ident == ATTR {
-                    if !matches!(result, Self::Normal) {
-                        panic!("Duplicate attribute: {}", ml.tokens);
-                    }
-
-                    if let Ok(ident) = syn::parse2::<syn::Ident>(ml.tokens.clone()) {
-                        if ident == "skip" {
-                            result = Self::Skip;
-                            continue;
-                        }
-                    }
-                    panic!("Invalid attribute: {}", ml.tokens);
-                }
-            }
-        }
-
-        result
-    }
-}
 
 /// Map field definitions to compressed field definitions.
 fn field_defs(fields: &syn::Fields) -> TokenStream2 {
@@ -212,61 +96,12 @@ fn construct_compressed(
         #compressed #construct_fields
     }
 }
-/// Construct original version of the struct from the compressed one
-fn construct_decompress(
-    // The original structure to construct, i.e. struct name or enum variant path
-    original: &TokenStream2,
-    variant: &synstructure::VariantInfo<'_>,
-) -> TokenStream2 {
-    let bound_fields: TokenStream2 = variant
-        .bindings()
-        .iter()
-        .map(|binding| {
-            let attrs = FieldAttrs::parse(&binding.ast().attrs);
-            let ty = &binding.ast().ty;
-            let cname = format_ident!("{}_c", binding.binding);
-
-            match attrs {
-                FieldAttrs::Skip => quote! {
-                    let #cname = Default::default();
-                },
-                FieldAttrs::Normal => {
-                    quote! {
-                        let #cname = <#ty as ::fuel_compression::DecompressibleBy<_, _>>::decompress(#binding, ctx).await?;
-                    }
-                }
-            }
-        })
-        .collect();
-
-    let construct_fields: TokenStream2 = variant
-        .bindings()
-        .iter()
-        .map(|binding| {
-            let cname = format_ident!("{}_c", binding.binding);
-            if let Some(fname) = &binding.ast().ident {
-                quote! { #fname: #cname, }
-            } else {
-                quote! { #cname, }
-            }
-        })
-        .collect();
-
-    let construct_fields = match variant.ast().fields {
-        syn::Fields::Named(_) => quote! {{ #construct_fields }},
-        syn::Fields::Unnamed(_) => quote! {(#construct_fields)},
-        syn::Fields::Unit => quote! {},
-    };
-
-    quote! {
-        #bound_fields
-        #original #construct_fields
-    }
-}
 
 /// Generate a match arm for each variant of the compressed structure
 /// using the given function to generate the pattern body.
-fn each_variant_compressed<F: FnMut(&synstructure::VariantInfo<'_>) -> TokenStream2>(
+pub fn each_variant_compressed<
+    F: FnMut(&synstructure::VariantInfo<'_>) -> TokenStream2,
+>(
     s: &synstructure::Structure,
     compressed_name: &TokenStream2,
     mut f: F,
@@ -295,21 +130,8 @@ fn each_variant_compressed<F: FnMut(&synstructure::VariantInfo<'_>) -> TokenStre
         .collect()
 }
 
-fn where_clause_push(w: &mut Option<syn::WhereClause>, p: TokenStream2) {
-    if w.is_none() {
-        *w = Some(syn::WhereClause {
-            where_token: syn::Token![where](proc_macro2::Span::call_site()),
-            predicates: Default::default(),
-        });
-    }
-    w.as_mut()
-        .unwrap()
-        .predicates
-        .push(syn::parse_quote! { #p });
-}
-
-/// Derives `Compressed` trait for the given `struct` or `enum`.
-pub fn compressed_derive(mut s: synstructure::Structure) -> TokenStream2 {
+/// Derives `Compress` trait for the given `struct` or `enum`.
+pub fn derive(mut s: synstructure::Structure) -> TokenStream2 {
     s.add_bounds(synstructure::AddBounds::None)
         .underscore_const(true);
 
@@ -433,17 +255,6 @@ pub fn compressed_derive(mut s: synstructure::Structure) -> TokenStream2 {
         construct_compressed(&construct, variant)
     });
 
-    let decompress_per_variant =
-        each_variant_compressed(&s, &quote! {#compressed_name}, |variant| {
-            let vname = variant.ast().ident.clone();
-            let construct = match &s.ast().data {
-                syn::Data::Struct(_) => quote! { #name },
-                syn::Data::Enum(_) => quote! {#name :: #vname },
-                syn::Data::Union(_) => unreachable!(),
-            };
-            construct_decompress(&construct, variant)
-        });
-
     let impls = s.gen_impl(quote! {
         gen impl ::fuel_compression::Compressible for @Self #w_impl {
             type Compressed = #compressed_name #g;
@@ -452,11 +263,6 @@ pub fn compressed_derive(mut s: synstructure::Structure) -> TokenStream2 {
         gen impl<Ctx, E> ::fuel_compression::CompressibleBy<Ctx, E> for @Self #w_impl_field_bounds_compress {
             async fn compress(&self, ctx: &mut Ctx) -> Result<Self::Compressed, E> {
                 Ok(match self { #compress_per_variant })
-            }
-        }
-        gen impl<Ctx, E> ::fuel_compression::DecompressibleBy<Ctx, E> for @Self #w_impl_field_bounds_decompress {
-            async fn decompress(compressed: &Self::Compressed, ctx: &Ctx) -> Result<Self, E> {
-                Ok(match compressed { #decompress_per_variant })
             }
         }
     });
