+++ fuel-tx/src/transaction/types/blob.rs
@@ -47,7 +47,10 @@ pub struct BlobMetadata;
 /// The body of the [`Blob`] transaction.
 #[derive(Clone, Default, Derivative)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "da-compression", derive(fuel_compression::Compressed))]
+#[cfg_attr(
+    feature = "da-compression",
+    derive(fuel_compression::Compress, fuel_compression::Decompress)
+)]
 #[derive(fuel_types::canonical::Deserialize, fuel_types::canonical::Serialize)]
 #[canonical(prefix = TransactionRepr::Blob)]
 #[derivative(Eq, PartialEq, Hash, Debug)]
